Refactor tests
This splits the overly large test file into multiple files, divided by
functional area of the tether, cmd, attach, and system for setting up base
system mechanisms.
Adds negative attach test for invalid session id.
